refactor(deploy): split retired adapter module
